ai_app/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
pymupdf>=1.24.0
google-generativeai>=0.8.0
httpx>=0.27.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

transformers>=4.30.0
torch>=2.0.0
